Coverity Scan Fixes (#2191)
* Metadata representation of resumed

* VirtualInput1, no temp

* LXC - Check sscanf result first

* protect observer count

* non const rval ref, move

* buffer not null terminated

* move instead of copy settings

* Timed race condition

* explicit to 0

* move in update accessor

* use ref instead of copy

* statm parse

* process.h scan check

* add a ref

* use localtime_r

* size_t for loops instead...

* uint16_t loops

* test threadpool, size t loops

* systeminfo test loop

* hash loop uint16_t

* Textfragment uint32_t

* asn1 string size

* const ref locals

* scnux64

* temp in probe

* move library

* return vals

* probe null wrong check

* moves, safety checks

* same as above

* copilot comments

* cov error fix

+++ Tests/unit/core/test_systeminfo.cpp
@@ -152,7 +152,7 @@ namespace Core {
         }
 
         float nonIdle = 0, VARIABLE_IS_NOT_USED idle = 0;
-        for (uint8_t index = 0; index < usageData.size(); ++index) {
+        for (size_t index = 0; index < usageData.size(); ++index) {
             if ((index == 3) || (index == 4)) {
                 idle += usageData[index];
             } else {
